mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.116.1",
    "mcp>=1.12.2",
    "openai>=1.97.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "sqlmodel>=0.0.24",
]
